# v7: reflection-based API