--- a/v7/rollout_integration_test.go
+++ b/v7/rollout_integration_test.go
@@ -4,11 +4,14 @@ import (
 	"bufio"
 	"context"
 	"encoding/csv"
+	"fmt"
 	"io"
 	"log"
 	"os"
 	"path/filepath"
+	"sort"
 	"strconv"
+	"strings"
 	"testing"
 )
 
@@ -21,27 +24,6 @@ type integrationTest struct {
 	sdkKey   string
 	fileName string
 	kind     int
-}
-
-func BenchmarkGetValue(b *testing.B) {
-	b.ReportAllocs()
-	logger := DefaultLogger(LogLevelError)
-	client := NewCustomClient(Config{
-		SDKKey:      integrationTests[0].sdkKey,
-		Logger:      logger,
-		RefreshMode: Manual,
-	})
-	client.Refresh(context.Background())
-	defer client.Close()
-	user := NewUser("unknown-identifier")
-	val := client.Bool("bool30TrueAdvancedRules", true, user)
-	if val != false {
-		b.Fatalf("unexpected result %#v", val)
-	}
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		client.Bool("bool30TrueAdvancedRules", true, user)
-	}
 }
 
 var integrationTests = []integrationTest{{
@@ -83,11 +65,15 @@ func (test integrationTest) runTest(t *testing.T) {
 		srv.setResponse(configResponse{body: contentForIntegrationTestKey(test.sdkKey)})
 		cfg = srv.config()
 	}
-	cfg.Logger = newTestLogger(t, LogLevelError)
+	tlogger := newTestLogger(t, LogLevelDebug).(*testLogger)
+	cfg.Logger = tlogger
 	cfg.SDKKey = test.sdkKey
 	client := NewCustomClient(cfg)
-	client.Refresh(context.Background())
 	defer client.Close()
+	err := client.Refresh(context.Background())
+	if err != nil {
+		t.Fatalf("cannot refresh: %v", err)
+	}
 
 	file, fileErr := os.Open(filepath.Join("../resources", test.fileName))
 	if fileErr != nil {
@@ -111,53 +97,60 @@ func (test integrationTest) runTest(t *testing.T) {
 			log.Fatal(err)
 		}
 		lineNumber++
-		var user *User
-		if line[0] != "##null##" {
-			identifier := line[0]
-			email := nullStr(line[1])
-			country := nullStr(line[2])
-			custom := map[string]string{}
-			if s := nullStr(line[3]); s != "" {
-				custom[customKey] = line[3]
-			}
-			user = NewUserWithAdditionalAttributes(identifier, email, country, custom)
-		}
-
-		for i, settingKey := range settingKeys {
-			var val interface{}
-			switch test.kind {
-			case valueKind:
-				val = client.getValue(settingKey, user)
-			case variationKind:
-				val = client.VariationID(settingKey, user)
-			default:
-				t.Fatalf("unexpected kind %v", test.kind)
-			}
-			expected := line[i+4]
-			var expectedVal interface{}
-			var err error
-			switch val := val.(type) {
-			case bool:
-				expectedVal, err = strconv.ParseBool(expected)
-			case int:
-				expectedVal, err = strconv.Atoi(expected)
-			case float64:
-				expectedVal, err = strconv.ParseFloat(expected, 64)
-			case string:
-				expectedVal = expected
-			default:
-				t.Fatalf("Value was not handled %#v", val)
-			}
-			if err != nil {
-				t.Fatalf("cannot parse expected value %q as %T: %v", expected, val, err)
-			}
-			if val != expectedVal {
-				t.Errorf("unexpected result for key %s at %s:%d; got %#v want %#v", settingKey, file.Name(), lineNumber, val, expectedVal)
-				for key, val := range user.attributes {
-					t.Logf("user %s: %q", key, val)
+		t.Run(fmt.Sprintf("line-%d", lineNumber), func(t *testing.T) {
+			var user User
+			if line[0] != "##null##" {
+				userVal := &UserValue{
+					Identifier: nullStr(line[0]),
+					Email:      nullStr(line[1]),
+					Country:    nullStr(line[2]),
 				}
+				if s := nullStr(line[3]); s != "" {
+					userVal.Custom = map[string]string{
+						customKey: s,
+					}
+				}
+				user = userVal
 			}
-		}
+			t.Logf("user %#v", user)
+
+			for i, settingKey := range settingKeys {
+				t.Run(fmt.Sprintf("key-%s", settingKey), func(t *testing.T) {
+					t.Logf("rule:\n%s", describeRules(client.fetcher.current(), settingKey))
+					tlogger.t = t
+					var val interface{}
+					switch test.kind {
+					case valueKind:
+						val = client.getValue(settingKey, user)
+					case variationKind:
+						val = client.VariationID(settingKey, user)
+					default:
+						t.Fatalf("unexpected kind %v", test.kind)
+					}
+					expected := line[i+4]
+					var expectedVal interface{}
+					var err error
+					switch val := val.(type) {
+					case bool:
+						expectedVal, err = strconv.ParseBool(expected)
+					case int:
+						expectedVal, err = strconv.Atoi(expected)
+					case float64:
+						expectedVal, err = strconv.ParseFloat(expected, 64)
+					case string:
+						expectedVal = expected
+					default:
+						t.Fatalf("value was not handled %T %#v; expected %q", val, val, expected)
+					}
+					if err != nil {
+						t.Fatalf("cannot parse expected value %q as %T: %v", expected, val, err)
+					}
+					if val != expectedVal {
+						t.Errorf("unexpected result for key %s at %s:%d; got %#v want %#v", settingKey, file.Name(), lineNumber, val, expectedVal)
+					}
+				})
+			}
+		})
 	}
 }
 
@@ -166,4 +159,41 @@ func nullStr(s string) string {
 		return ""
 	}
 	return s
+}
+
+func describeRules(cfg *config, specificKey string) string {
+	if cfg == nil {
+		return "no config"
+	}
+
+	var buf strings.Builder
+	printf := func(f string, a ...interface{}) {
+		fmt.Fprintf(&buf, f, a...)
+	}
+	printResult := func(value interface{}, variationID string) {
+		printf("\t→ %T %#v; variation %q\n", value, value, variationID)
+	}
+	keys := make([]string, 0, len(cfg.root.Entries))
+	for key := range cfg.root.Entries {
+		keys = append(keys, key)
+	}
+	sort.Strings(keys)
+	for _, key := range keys {
+		if specificKey != "" && key != specificKey {
+			continue
+		}
+		entry := cfg.root.Entries[key]
+		printf("%q\n", key)
+		for _, rule := range entry.RolloutRules {
+			printf("\t: user.%s %v %q\n", rule.ComparisonAttribute, rule.Comparator, rule.ComparisonValue)
+			printResult(rule.Value, rule.VariationID)
+		}
+		for _, rule := range entry.PercentageRules {
+			printf("\t: %d%%\n", rule.Percentage)
+			printResult(rule.Value, rule.VariationID)
+		}
+		printf("\t: default\n")
+		printResult(entry.Value, entry.VariationID)
+	}
+	return buf.String()
 }