This allows a user-defined User object to be passed in,
making it easier to use an arbitrary domain object for the
user data and avoiding the need to convert to and from the
intermediate string format for numeric comparisons.

Although the specification is simple (allow any numeric type or any
type with a String method), that allows for quite a number of possible
combinations, so the eval code becomes somewhat more complex.
It might be better to restrict the number of possible types
(for example, by only allowing `int` and `float64` and `string`)
or use floating point comparisons always for the numeric operators,
which might reduce the amount of code required.
I can't quite decide!

Benchmarks compared with v6:

```
name                      old time/op    new time/op    delta
Get/one-of-8                 339ns ± 4%     167ns ±18%  -50.73%  (p=0.008 n=5+5)
Get/less-than-with-int-8     436ns ± 5%     106ns ± 6%  -75.76%  (p=0.008 n=5+5)
Get/with-percentage-8        678ns ±16%     469ns ± 2%  -30.86%  (p=0.008 n=5+5)

name                      old alloc/op   new alloc/op   delta
Get/one-of-8                  360B ± 0%       64B ± 0%  -82.22%  (p=0.008 n=5+5)
Get/less-than-with-int-8      360B ± 0%        8B ± 0%  -97.78%  (p=0.008 n=5+5)
Get/with-percentage-8         456B ± 0%      136B ± 0%  -70.18%  (p=0.008 n=5+5)

name                      old allocs/op  new allocs/op  delta
Get/one-of-8                  3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.008 n=5+5)
Get/less-than-with-int-8      3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.008 n=5+5)
Get/with-percentage-8         5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.008 n=5+5)
```